+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## [0.5.0 - (2019-10-14)](https://github.com/InterConnecta/zoho-crm/compare/0.4.1...0.5.0)
+
+### Added
+  - [get-modules] Added getModules method ([6c86de0](https://github.com/InterConnecta/zoho-crm/commit/6c86de0))
+  - getRecords ([34e9760](https://github.com/InterConnecta/zoho-crm/commit/34e9760))
+  - [get-attachments] Added getAttachments method ([3634133](https://github.com/InterConnecta/zoho-crm/commit/3634133))
+  - [get-attachments] Added getAttachmentData util function ([34c3e61](https://github.com/InterConnecta/zoho-crm/commit/34c3e61))
+
+## [0.4.1 - (2019-10-08)](https://github.com/InterConnecta/zoho-crm/compare/0.4.0...0.4.1)
+
+### Added
+  - [related-records] Get line items in Related records ([4c6d162](https://github.com/InterConnecta/zoho-crm/commit/4c6d162))
+
+
+## [0.4.0 - (2019-10-08)](https://github.com/InterConnecta/zoho-crm/compare/0.3.0...0.4.0)
+
+### Added
+  - [create-records] Added createRecords method ([5892bb4](https://github.com/InterConnecta/zoho-crm/commit/5892bb4))
+
+
 ## [0.3.0 - (2019-10-03)](https://github.com/InterConnecta/zoho-crm/compare/0.2.0...0.3.0)
 
 ### Added